+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.4.1] - 2025-09-01
+
+### Fixed
+- `is_created_suspended` falsely returning `false` due to thread hijacking performed by the Steam Overlay
+- Thread handle leakage in `iter_threads`
+
 ## [v0.4.0] - 2025-08-29
 
 ### Breaking Changes